pyproject.toml
[project]
name = "monospace-blog"
version = "1.0.0"
description = "A static blog generator that converts markdown files with YAML frontmatter into beautiful HTML with a monospace theme"
authors = [
    {name = "Christopher C. Smith", email = "christopher.smith@promptlytechnologies.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.2",
    "PyYAML>=6.0.1",
    "python-frontmatter>=1.0.0",
    "Pillow>=10.1.0",
    "mistletoe>=1.4.0",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://christophercarrollsmith/monospace-blog"
Repository = "https://github.com/chriscarrollsmith/monospace-blog"

[tool.uv]
dev-dependencies = [
    "playwright>=1.53.0",
] 
